Add files via upload
The task is to perform 400 simple linear regression models to predict movie ratings, using the ratings of the other 399 movies as predictors. For each movie, we will find the predictor movie that predicts its ratings the best and report the average COD of those 400 models, along with a histogram of the COD values and a table with the 10 easiest and 10 hardest movies to predict, their associated COD values, and the best predictor movie.